% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Routines.R
\name{Assimilation}
\alias{Assimilation}
\title{Assimilation routine as in PCSE/WOFOST}
\usage{
Assimilation(AMAXTB, TMPFTB, KDIFTB, EFFTB, TMNFTB, d, lai, sgd, dp, sinbm,
  sinld, cosld, dvs, tday, w, t, tmins)
}
\description{
Computes "pgass" (Potential assimilation rate). Calls functions "totass()"
and "assim()"
}
