% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuxillaryFunctions.R
\name{combineObj}
\alias{combineObj}
\title{Combine weather object}
\usage{
combineObj(obj)
}
\arguments{
\item{obj}{List of weatherObjects}
}
\description{
Returns a stiched WeatherObject from a list of consecutive WeatherObjects.
}
