% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\docType{methods}
\name{CropObject,list-method}
\alias{CropObject,list-method}
\title{S4 Method for generic "CropObject()", list}
\usage{
\S4method{CropObject}{list}(...)
}
\arguments{
\item{...}{Named list where each named element corresponds to an omonimous
slot in the CropObject to be created.}
}
\description{
S4 Method for generic "CropObject()", list
}
