% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\name{CropObject}
\alias{CropObject}
\title{S4 Class "CropObject"}
\usage{
CropObject(...)
}
\arguments{
\item{...}{Named list where each named element corresponds to an omonimous
slot in the CropObject to be created. Alternatively slots can be filled
individually (e.g. with "CropObject(CVR = 2, SPA = 7)").}
}
\description{
Contains crop parameters
}
