% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{cropVarList}
\alias{cropVarList}
\title{List of available crops and varieties.}
\format{An object of class \code{data.frame} with 56 rows and 2 columns.}
\usage{
cropVarList
}
\description{
Dataframe of two columns describing all available crops and varieties on
https://github.com/ajwdewit/WOFOST_crop_parameters .
}
\keyword{datasets}
