% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuxillaryFunctions.R
\name{dwn.crop}
\alias{dwn.crop}
\title{Download crop parameters}
\usage{
dwn.crop(cropName = "potato", variety = "Potato_701")
}
\arguments{
\item{cropName}{Character identifying the name of the crop to be
downlowaded. See list on link above.}

\item{variety}{Character.The variety name as mentioned in they yaml file
on link above.}
}
\value{
CropObject
}
\description{
Downloads crop parameter files from
"https://github.com/ajwdewit/WOFOST_crop_parameters.git"
}
\examples{
crop <- dwn.crop(cropName='potato', variety='Potato_701')
crop

}
