% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Routines.R
\name{Evapotranspiration}
\alias{Evapotranspiration}
\title{Evapotranspiration routine}
\usage{
Evapotranspiration(dvs, w, lai, sm, t, dsos, SM0, DEPNR, SMFCF, IAIRDU,
  IOX, CFET, SMW, CRAIRC, KDIFTB)
}
\arguments{
\item{dvs}{Numeric. Development stage}

\item{w}{Weather table}

\item{lai}{Numeric. Leaf Area Index LAI}

\item{sm}{Numeric. Volumetric soil moisture content}

\item{t}{Integar. Time step of the model}

\item{dsos}{Days since oxygen stress, computed in classic waterbalance,
accumulates the number of consecutive days of oxygen stress}

\item{DEPNR}{Crop group number}

\item{SMFCF}{Soil moisture content at field capacity [cm3/cm3]}

\item{IAIRDU}{Indicates presence (1) or absence (0) of airducts in the plant}

\item{IOX}{Flag; controlling calculation of potential (0) or
water limited yield (1)}

\item{CFET}{Correction parameter of potential evapotranspiration rate}

\item{SMW}{Soil moisture content at wilting point}

\item{CRAIRC}{Critical air content for root aeration}

\item{KDIFTB}{Extinction coefficient for diffuse visible radiation as
function of DVS}
}
\description{
Routine Evapotranspiration computes evapotranspiration rates from crop
and soil input parameters
}
