% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{exampleManager}
\alias{exampleManager}
\title{Example of ManagerObject.}
\format{An object of class \code{ManagerObject} of length 1.}
\usage{
exampleManager
}
\description{
Example of object of ManagerObject class.
}
\keyword{datasets}
