% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuxillaryFunctions.R
\name{get.wofost.results}
\alias{get.wofost.results}
\title{Extract test file variables}
\usage{
get.wofost.results(direc, name, variable, section = "results")
}
\arguments{
\item{direc}{Directory where the relevant yaml file is saved}

\item{name}{Name of test file (including extension)}

\item{variable}{Name of variable to retrieve}

\item{section}{Section of yaml file where variable is stored}
}
\description{
Extracts any variable from yaml test files.
}
