% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growingWindow.R
\name{growingWindow}
\alias{growingWindow}
\title{Find growing window}
\usage{
growingWindow(crop = "potato", variety = "Potato_701", w,
  allVar = FALSE)
}
\arguments{
\item{crop}{Character identifying the name of the crop to be
downlowaded with function dwn.crop().}

\item{variety}{Character.The variety name. See omonimous variable in
dwn.crop().}

\item{w}{WeatherObject encompassing the testing period. Should end with
one extra year to allow for testing the sowing of december of second-to-last
year.}

\item{allVar}{Logical. If TRUE all crop-variety combinations available
on "https://github.com/ajwdewit/WOFOST_crop_parameters.git" will be run.}
}
\description{
Determins if a crop can reach maturity when grown in a specific location
}
