% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuxillaryFunctions.R
\name{load.crop}
\alias{load.crop}
\title{Load crop parameters}
\usage{
load.crop(cropName = "potato", variety = "Potato_701", crLocal)
}
\arguments{
\item{cropName}{Character identifying the name of the crop to be
downlowaded. See list on link above.}

\item{variety}{Character.The variety name as mentioned in they yaml file
on link above.}

\item{crLocal}{Character vector. Path where all crop yaml files are stored.}
}
\value{
CropObject
}
\description{
Load crop parameter files from locally stored yaml files
}
