% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuxillaryFunctions.R
\name{make.afgen}
\alias{make.afgen}
\title{Afgen vector to table}
\usage{
make.afgen(tab)
}
\arguments{
\item{tab}{Afgen table expressed as vector.}
}
\description{
Renders afgen tables expressed as "vectors" into a "matrix" format.
}
