% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Penman.R
\name{Penman}
\alias{Penman}
\title{Evapo(transpiration) following Penman's formula}
\usage{
Penman(w)
}
\arguments{
\item{w}{WeatherObject where E0, ES0 and ET0 are not computed yet.}
}
\value{
WeatherObject w with added E0, ES0 and ET0 slots.
}
\description{
Computes and adds E0, ES0, ET0 to a WeatherObject where these are absent
using Penmans formulation.
}
\details{
This routine calculates the potential evapo(transpi)ration rates
from a free water surface (E0), a bare soil surface (ES0), and a crop canopy
(ET0) in mm/d. For these calculations the analysis by Penman is followed
(Frere and Popov, 1979;Penman, 1948, 1956, and 1963).
}
