% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{penrynWeather}
\alias{penrynWeather}
\title{Temporary weather file from Penryn for playing around.}
\format{An object of class \code{WeatherObject} of length 1.}
\usage{
penrynWeather
}
\description{
Temporary weather file from Penryn for playing around.
}
\keyword{datasets}
