% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\docType{methods}
\name{plot,SimulationObject-method}
\alias{plot,SimulationObject-method}
\title{S4 Method for generic "plot()", SimulationObject}
\usage{
\S4method{plot}{SimulationObject}(x, var = NULL)
}
\arguments{
\item{x}{SimulationObject S4 object}

\item{var}{Character vector of variable names that will be plotted.}
}
\description{
Plots an overview of the output of function Wofost
}
\examples{
# Generate SimulationObjects
w <- randomWeatherLong
s <- randomSoil
cr <- dwn.crop()
wopp <- Wofost(w, cr)
womn <- Wofost(w, cr, manager = exampleManager)

# Plot SimulationObjects
plot(wopp)
plot(wopp, var = c('dvs', 'tagp'))

plot(womn)
plot(womn, var = 'tagp')

}
