% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\docType{methods}
\name{plot,WeatherObject-method}
\alias{plot,WeatherObject-method}
\title{S4 Method for generic "plot()", WeatherObject}
\usage{
\S4method{plot}{WeatherObject}(x, var = NULL)
}
\arguments{
\item{x}{WeatherObject S4 object}

\item{var}{Character vector of variable names that will be plotted.}
}
\description{
Plots an overview of WeatherObjects
}
\examples{
w <- randomWeatherLong
plot(w)
plot(w, var = c('TEMP', 'RAIN'))

}
