% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuxillaryFunctions.R
\name{plot.wofost.test}
\alias{plot.wofost.test}
\title{Plot Wofost test results}
\usage{
\method{plot}{wofost.test}(output, control, weather, precision = FALSE,
  keypress = TRUE)
}
\arguments{
\item{output}{Dataframe of variables outputted by the function wofots.test.
Column order must be the same as in "control".}

\item{control}{Control section of yaml test file as returned by function
make.afgen}

\item{weather}{wather section of yaml test file as returned by function
make.afgen}

\item{precision}{Precision section of yaml test file as returned by function
make.afgen}

\item{keypress}{Logical. If TRUE, plots one graph at the time after a key is
pressed. if FALSE, returns them all together, par(mfrow())
can be used to place all plots in the same image.}
}
\description{
Plots graphs of outputs of test trials
}
