% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuxillaryFunctions.R
\name{plotWofost}
\alias{plotWofost}
\title{Plots the output of Wofost()}
\usage{
plotWofost(out)
}
\arguments{
\item{out}{output of Wofost().}
}
\value{
Single plot
}
\description{
Plots each output variable of function Wofost() in a single image
}
