% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Test_PotentialProduction.R
\name{Potential_production}
\alias{Potential_production}
\title{Wofost Potential Production}
\usage{
Potential_production(crop, w, lat, ext, waterlimited = F)
}
\description{
Computes potential production as per Wofost crop growth algorithm
}
