% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{randomSoil}
\alias{randomSoil}
\title{Example SoilObject}
\format{An object of class \code{SoilObject} of length 1.}
\usage{
randomSoil
}
\description{
Example of object of SoilObject class.
}
\keyword{datasets}
