% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{randomWeather}
\alias{randomWeather}
\title{Example weatherObject}
\format{An object of class \code{WeatherObject} of length 1.}
\usage{
randomWeather
}
\description{
Example of object of weatherObject class.
}
\keyword{datasets}
