% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{randomWeatherLong}
\alias{randomWeatherLong}
\title{Example of multi-year WeatherObject.}
\format{An object of class \code{WeatherObject} of length 1.}
\usage{
randomWeatherLong
}
\description{
Example of muli-year object of WeatherObject class.
Useful for using with a ManagerObject.
}
\keyword{datasets}
