% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuxillaryFunctions.R
\name{read.test.yaml}
\alias{read.test.yaml}
\title{Reads .yaml test file}
\usage{
read.test.yaml(yamlFile)
}
\arguments{
\item{yamlFile}{Path to the .yaml file or list object returned by function
yaml::read_yaml.}
}
\value{
List version of yaml file.
}
\description{
Reads .yaml test files published in De Wit et al. 2019,
"25 years of the WOFOST cropping systems model", and returns R list.
}
