% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuxillaryFunctions.R
\name{read.wofost.yaml}
\alias{read.wofost.yaml}
\title{Reads .yaml test file}
\usage{
read.wofost.yaml(yamlFile)
}
\arguments{
\item{yamlFile}{Path to the .yaml file}
}
\description{
Reads .yaml test files published in De Wit et al. 2019,
"25 years of the WOFOST cropping systems model", and returns R list.
}
