% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RootDynamicsTest.R
\name{RootDynamics}
\alias{RootDynamics}
\title{Roots Dynamics for testing}
\usage{
RootDynamics(RDI, RDMCR, RDMSOL, RDRRTB, RRI, TDWI, w, ext)
}
\description{
Function that performs the root-dynamics-relevant code in isolation.
Used for testing the roots dynamics routine with the testing data of
De Wit et al. 2018
}
