% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\docType{methods}
\name{show,ManagerObject-method}
\alias{show,ManagerObject-method}
\title{S4 Method for generic "show()", ManagerObject}
\usage{
\S4method{show}{ManagerObject}(object)
}
\arguments{
\item{object}{ManagerObject S4 object}
}
\description{
Prints a summary of the manager object when typed or when
"show(object_name)" is called
}
