% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\docType{methods}
\name{show,SimulationObject-method}
\alias{show,SimulationObject-method}
\title{S4 Method for generic "show()"}
\usage{
\S4method{show}{SimulationObject}(object)
}
\description{
Prints a summary of the simulation object when typed or when
"show(object_name)" is called
}
