% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\docType{methods}
\name{show,SoilObject-method}
\alias{show,SoilObject-method}
\title{S4 Method for generic "show()", SoilObject}
\usage{
\S4method{show}{SoilObject}(object)
}
\arguments{
\item{object}{CropObject S4 object}
}
\description{
Prints a summary of the simulation object when typed or when
"show(object_name)" is called
}
