% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\docType{methods}
\name{show,WeatherObject-method}
\alias{show,WeatherObject-method}
\title{S4 Method for generic "show()", WeatherObject}
\usage{
\S4method{show}{WeatherObject}(object)
}
\arguments{
\item{object}{CropObject S4 object}
}
\description{
Prints a summary of the weather object when typed or when
"show(object_name)" is called
}
