% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\docType{class}
\name{SimulationObject-class}
\alias{SimulationObject-class}
\alias{SimulationObject}
\title{S4 Class "SimulationObject"}
\description{
Contains crop parameters, rates and states for the current time step in
the simulation.
}
