% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\docType{methods}
\name{SoilObject,list-method}
\alias{SoilObject,list-method}
\title{S4 Method for generic "SoilObject()", list}
\usage{
\S4method{SoilObject}{list}(...)
}
\arguments{
\item{...}{Named list where each named element corresponds to an omonimous
slot in the SoilObject to be created.}
}
\description{
S4 Method for generic "SoilObject()", list
}
