% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\name{SoilObject}
\alias{SoilObject}
\title{S4 Class "SoilObject"}
\usage{
SoilObject(...)
}
\arguments{
\item{...}{Named list where each named element corresponds to an omonimous
slot in the SoilObject to be created. Alternatively slots can be filled
individually (e.g. with "SoilObject(CVR = 2, SPA = 7)").}
}
\description{
Contains soil parameters
}
