% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuxillaryFunctions.R
\name{subsetObj}
\alias{subsetObj}
\title{Subsets weather object}
\usage{
subsetObj(obj, interval)
}
\arguments{
\item{interval}{Either a vector of integers representing the indices of the
days to extract or a vector of class="Date" and length=2 containing the
initial and final date of the period to extract.}
}
\value{
WeatherObject of the selected period
}
\description{
Returns a section of the WeatherObject based on dates
}
