% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuxillaryFunctions.R
\name{test.afgen}
\alias{test.afgen}
\title{Tests afgen() function}
\usage{
test.afgen(at)
}
\arguments{
\item{at}{atrix. Afgen table}
}
\description{
Performs a graphical test to check that afgen() function performs
correcly. Also useful to visualise the shape of the function represented
by an afgen table.
}
