% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tests.R
\name{test.Evapotranspiration}
\alias{test.Evapotranspiration}
\title{Evapotranspiration routine testing}
\usage{
test.Evapotranspiration(yamlFile, plt = TRUE, ...)
}
\arguments{
\item{yamlFile}{Path to yaml test file}

\item{plt}{Logical. If TRUE prints plots of results with plot.wofost.test}

\item{...}{Arguments passed to function plot.wofost.test}
}
\value{
List of 2 elements. 'output': the variables values produced by the
        script 'control': the variables values saved in the yaml file.
}
\description{
Tests the Evapotranspiration routine with the testing data of
De Wit et al. 2018
}
