% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuxillaryFunctions.R
\name{test}
\alias{test}
\title{Test Wofost}
\usage{
test(component = "All", complete = FALSE)
}
\arguments{
\item{component}{Character. Name of module to test among:
"astro","phenology","transpiration","potentialproduction",
"waterlimitedproduction".

If not specified will run all available components.}

\item{complete}{Logical. If FALSE only a subset of 3 files per set will be
used for testing.}
}
\description{
Tests different Wofost modules by comparison with the testing sets
published by De Wit et al. (2018).
}
\details{
Although testing sets for more modules have been made available in
De Wit et al. 2018, it does not make sense to test some available modules
since they have not been implemente in WofostR as independent components,
but rather they have been hard-coded into other functions.
Examples of these are 'partitioning', 'root dynamics', 'leaf dynamics',
and others that do not exhist as independet R functions.

"phenology" tests the stand-alone module
Phenology() used by function growingWindow(). Phenological processes
are hardcoded within in Wofost() and can only be tested through
"potentialproduction" and "waterlimitedproduction".
}
