% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Routines.R
\name{Totas_Assim}
\alias{Totas_Assim}
\title{Totas & Assim}
\usage{
Totas_Assim(tday, d, sd, lat, sgd, sinbm, dp, s = 0.2, lai, dvs, t, tlow,
  KDIFTB, AMAXTB, TMPFTB, TMNFTB, EFFTB)
}
\arguments{
\item{tday}{Average daytime temperature}

\item{d}{Astronomical day length}

\item{sd}{Solar declination}

\item{lat}{Latitude}

\item{sgd}{Daily global radiation}

\item{sinbm}{Corrected integral of solar height over the day (4.28)}

\item{dp}{Diffuse radiation perpendicular to the direction light (4.31)}

\item{s}{Scattering coefficient fraction (transmission and reflection) (=0.2?)}

\item{lai}{Leaf area index of whole canopy}

\item{dvs}{Development stage}

\item{t}{Day in the loop}

\item{AMAXTB}{Afgen table (inst. gross assimilation rate at light saturation
vs. development stage)}

\item{TMPFTB}{Afgen table (inst. gross assimilation rate at light saturation
vs. daytime temperature)}

\item{TMNFTB}{Afgen table (inst. gross assimilation rate at light saturation
vs. 7 days running minumum temperature)}

\item{EFFTB}{Afgen table (light use efficiency vs. daily mean temp),
crop-specific [kg ha-1 hr-1 j-1 m2 s]}

\item{kdf}{Extinction coefficient for the diffuse radiation flux}
}
\description{
Routines Total and Assim combined. Computes total daily gross CO2
assimilation rate for the whole canopy
}
