% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\docType{methods}
\name{WeatherObject,list-method}
\alias{WeatherObject,list-method}
\title{S4 Method for generic "WeatherObject()", list}
\usage{
\S4method{WeatherObject}{list}(...)
}
\description{
S4 Method for generic "WeatherObject()", list
}
