% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\name{WeatherObject}
\alias{WeatherObject}
\title{S4 Class "WeatherObject"}
\usage{
WeatherObject(...)
}
\arguments{
\item{...}{Named list where each named element corresponds to an omonimous
slot in the WeatherObject to be created. Alternatively slots can be filled
individually (e.g. with "WeaterObject(CVR = 2, SPA = 7)").}
}
\description{
Contains meteorological driving variables.
}
