% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winSelectCrops.R
\name{winSelectCrops}
\alias{winSelectCrops}
\title{Select crops have viable growing window}
\usage{
winSelectCrops(gw)
}
\arguments{
\item{gw}{Growing Window. The output of growingWindow() function.}
}
\description{
Takes the output of growingWindow() and returns a selection of
crop/varieties that reached maturity with their corresponding sowing
dates. This is useful for running Wofost on those with function
"winSelectWofost()".
}
