% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winSelectCrops.R
\name{winSelectWofost}
\alias{winSelectWofost}
\title{Runs wofost for winSelectCrops() output}
\usage{
winSelectWofost(gwOut, w, waterlimited = FALSE)
}
\arguments{
\item{gwOut}{Selected crops as outputted from winSelectCrops().}

\item{w}{WeatherObject. Must encompass the period from the earliest
sowing date to 730 days after the last sowing date.}

\item{waterlimited}{Logical. If FALSE (default) runs potential production.
If TRUE, free draining water-limited production.}
}
\description{
Takes the output of winSelectCrops() and runs a Wofost() run for each
sowing date for each selected crop.
}
