% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Routines.R
\name{Wofost}
\alias{Wofost}
\title{Wofost Potential Production}
\usage{
Wofost(crop, w, waterlimited = F, finishType = "maturity",
  startType = "sowing", varReturn = NULL)
}
\arguments{
\item{crop}{CorpObject. Object of class CropObject containing the specific
crop parameters. See ?CropObject.}

\item{w}{WeatherObject. Object of class CropObject containing the climatic
driving variables. See ?WeatherObject.}

\item{waterlimited}{Logical. If TRUE waterlimited production is modelled,
potential production is modelled by default (waterlimited = F)}

\item{finishType}{Variable describing the conditionst riggering
the end of the simulation.
Can be any one of the following:

"maturity": The model is terminated 7 days after maturity is reached
Integer [1:365]: Maximum number of days for which the model is run.}

\item{varReturn}{Character vector specifying which variables to output.
Potentially, any of the variables produced inside the Wofost function
can be returned. However the use of carReturn = NULL (default) is
encouraged. By default returning variables described in "Returns".}
}
\description{
Computes potential production as per Wofost crop growth algorithm
}
