% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Routines.R
\name{Astro}
\alias{Astro}
\title{Astro}
\usage{
Astro(w, t, lat)
}
\arguments{
\item{w}{Dataframe. Daily weather table}

\item{t}{Integer. Time steps (days) passed from emergence (dvs==0)}

\item{lat}{Numeric. Latitude}
}
\description{
Routine ASTRO calculates day length, some intermediate variables for the
calculation of the solar elevation, the integral of the solar elevation
over a day and the fraction of diffuse radiation.
}
