% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\docType{methods}
\name{CropObject,list-method}
\alias{CropObject,list-method}
\title{S4 Method for generic "CropObject()"}
\usage{
\S4method{CropObject}{list}(...)
}
\description{
S4 Method for generic "CropObject()"
}
