% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\name{CropObject}
\alias{CropObject}
\title{S4 Class "CropObject"}
\usage{
CropObject(...)
}
\arguments{
\item{...}{Named list where each named element corresponds to an omonimous
slot in the CropObject to be created. Alternatively slots can be filled
individually (e.g. with "CropObject(IDSL = 2, SPA = 7)").}
}
\description{
Contains crop parameters
}
\details{
Instances of this class are normally generated by calling
dwn.crop() or load.crop() which retrieve all crop parameters from complete
crop descriptions in the form of .yamls files.
}
\examples{
cr <- dwn.crop() # require internet connection
cr
str(cr, list.len = 3)

}
