% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growingWindow.R
\name{growingWindow}
\alias{growingWindow}
\title{Find growing window}
\usage{
growingWindow(crop = "potato", variety = "Potato_701", w,
  allVar = FALSE, crLocal = NULL, activate.verndvs = FALSE)
}
\arguments{
\item{crop}{Character identifying the name of the crop to be
downlowaded with function dwn.crop().}

\item{variety}{Character.The variety name. See omonimous variable in
dwn.crop().}

\item{w}{WeatherObject encompassing the testing period. Should end with
one extra year to allow for testing the sowing of december of second-to-last
year.}

\item{allVar}{Logical. If TRUE all crop-variety combinations available
on "https://github.com/ajwdewit/WOFOST_crop_parameters.git" or on the
local directory crLocal will be run.}

\item{crLocal}{Character vector. If crops are stored locally, this is
the path where all crop yaml files are stored.}

\item{activate.verndvs}{Logical. If FALSE (default), variable "verndvs"
is deactivated so that there is a hard limit for crops that do not meet
vernalisation requirements. If TRUE, crops are left to grow past a specific
developmnent stage also if vernalisation requirements are not fullfilled.
A warning is issued in this case.}
}
\description{
Determins if a crop can reach maturity when grown at a specific location
}
\examples{
# Run single crop
out <- growingWindow(w = randomWeatherLong,
                     crop = "potato",
                     variety = "Potato_701")

# Run all available crops
# Not run. Takes some minutes
# out <- growingWindow(w = randomWeatherLong,
#                      allVar = TRUE)

}
