% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phenology.R
\name{Phenology}
\alias{Phenology}
\title{Stand-alone phenology module}
\usage{
Phenology(crop, w, startType = "sowing", finishType = "maturity")
}
\arguments{
\item{crop}{CorpObject. Object of class CropObject containing the specific
crop parameters. See ?CropObject.}

\item{w}{WeatherObject. Object of class CropObject containing the climatic
driving variables. See ?WeatherObject.}

\item{startType}{Development stage at which the simulation is started.
Either "sowing" or "emergence".}

\item{finishType}{Variable describing the conditionst riggering
the end of the simulation.
Can be either "maturity" -The model is terminated 7 days after maturity is
reached - or
an integer [1:365] -Maximum number of days for which the model is run.}
}
\description{
Computes the phenological development of the crop
}
