% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\docType{methods}
\name{show,SimulationObject-method}
\alias{show,SimulationObject-method}
\title{S4 Method for generic "show()", SimulationObject}
\usage{
\S4method{show}{SimulationObject}(object)
}
\arguments{
\item{object}{SimulationObject S4 object}
}
\description{
Prints a summary of the manager object when typed or when
"show(object_name)" is called
}
