% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\docType{class}
\name{SimulationObject-class}
\alias{SimulationObject-class}
\alias{SimulationObject}
\title{S4 Class "SimulationObject"}
\arguments{
\item{description:}{Dataframe containing metadata of the simulation}

\item{variables:}{List of length equal to nrow("description") containing
the output variables. Each top level element of "variables" represents a
crop.}
}
\description{
Contains the output of function Wofost()
}
\examples{

}
