% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuxillaryFunctions.R
\name{test}
\alias{test}
\title{Test Wofost}
\usage{
test(component, complete = FALSE)
}
\arguments{
\item{component}{Character. Either one of the testing sets released
in De Wit 2018. Any of "assimilation", "astro", "leafdynamics",
"partitioning", "phenology", "potentialproduction", "respiration",
"rootdynamics", "transpiration", "waterlimitedproduction".
However, currently limited to PP, LD, A.}

\item{complete}{Logical. If FALSE only a subset of 3 files per set will be
used for testing.}
}
\description{
Tests different components of this R implementation of Wofost by
comparison with the testing sets published in
De Wit 2018 (Agricultural Systems journal).
}
