% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Routines.R
\name{totass}
\alias{totass}
\title{Function totass as in PCSE/WOFOST (Python version)}
\usage{
totass(d, amax, eff, lai, kdif, sgd, dp, sinbm, sinld, cosld)
}
\description{
This routine calculates the daily total gross CO2 assimilation by
performing a Gaussian integration over time. At three different times of
the day, irradiance is computed and used to calculate the instantaneous
canopy assimilation, whereafter integration takes place. More information
on this routine is given by Spitters et al. (1988).
}
