% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winSelectCrops.R
\name{winSelectWofost}
\alias{winSelectWofost}
\title{Runs wofost for winSelectCrops() output}
\usage{
winSelectWofost(gwOut, w, s, crLocal = NULL, waterLimited = FALSE,
  outDir = "")
}
\arguments{
\item{gwOut}{Selected crops as outputted from winSelectCrops().}

\item{w}{WeatherObject. Must encompass the period from the earliest
sowing date to 730 days after the last sowing date.}

\item{s}{SoilObject.}

\item{crLocal}{Character vector. If crops are stored locally, this is
the path where all crop yaml files are stored.}

\item{waterLimited}{Logical. If FALSE (default) runs potential production.
If TRUE, free draining water-limited production.}

\item{outDir}{Directory path (to be created) where the output is saved.}
}
\description{
Deprecated! This function has been superseeded by cropSuitWofost().
Takes the output of winSelectCrops() and runs a Wofost() run for each
sowing date for each selected crop.
}
