% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wofost.R
\name{Wofost}
\alias{Wofost}
\title{Wofost Wofost crop simulation}
\usage{
Wofost(w, crop = NULL, soil = NULL, manager = NULL,
  waterLimited = FALSE, startType = "sowing",
  finishType = "maturity", cropDir = NULL, varReturn = NULL)
}
\arguments{
\item{w}{WeatherObject. Object of class CropObject containing the climatic
driving variables. See ?WeatherObject.}

\item{crop}{CorpObject.
Object of class CropObject containing the specific
crop parameters. See ?CropObject. To specify a crop sequence (rotation)
use "manager".}

\item{soil}{SoilObject. Object of class SoilObject containing the soil
parameters. See ?SoilObject.}

\item{manager}{Optional ManagerObject. If present "startType" and
"finishType will be ignored".}

\item{waterLimited}{Logial. Wether to simulate potential production
("= FALSE", default), or water limited production ("= TRUE")}

\item{startType}{Development stage at which the simulation is started.
Either "sowing" or "emergence". Ignored if "manager" is specified.}

\item{finishType}{Variable describing the conditionst triggering
the end of the simulation.
Can be either "maturity" -The model is terminated 7 days after maturity is
reached - or
an integer [1:365] -Maximum number of days for which the model is run.
Ignored if "manager" is specified.}

\item{cropDir}{Character vector containing the directory where crop files
are stored as .yaml files. If not specified and a CropObject is not
assigned to parameter "crop", crop files will be downloaded from
"https://github.com/ajwdewit/WOFOST_crop_parameters".}

\item{varReturn}{Character vector specifying which variables to output.
Potentially, any of the variables produced inside the Wofost function
can be returned. However the use of carReturn = NULL (default) is
encouraged. By default returning variables described in "Returns".}
}
\description{
Computes potential or water-limited production on freely draining soils
without water table.
}
\examples{

# Run Potential Production
cr <- dwn.crop()
outPP <- Wofost(crop = cr, w = randomWeather)
plot(outPP)

# Run Water-Limited Production
cr <- dwn.crop()
outFD <- Wofost(crop = cr, w = randomWeather, soil = randomSoil,
                waterLimited = TRUE)
plot(outFD)

# Run crop rotation with ManagerObject
outMN <- Wofost(w = randomWeatherLong, manager = exampleManager)
plot(outMN)


}
